ticket controller added required apis

+++ src/main/java/ba/unsa/pmf/pragma/db/repository/TicketTypeRepository.java
@@ -3,5 +3,5 @@ package ba.unsa.pmf.pragma.db.repository;
 import ba.unsa.pmf.pragma.db.entity.TicketType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketTypeRepository extends JpaRepository<TicketType,Long> {
+public interface TicketTypeRepository extends JpaRepository<TicketType,Short> {
 }
